Change 'modelica_integer' to 'int'. (#9064)
* Change modelica_integer to int

* Change _index_t to modelica_integer (=int)

* Change modelica_integer to int in OpenCL code.

* Update expected outputs for test cases.

  - These are due to the fact that we now use the limit `INT_MAX/2`
    instead of `LONG_MAX/2` for integer maximums.

* Add the test case from the ticket to the testsuite.

+++ OMCompiler/SimulationRuntime/c/meta/meta_modelica_builtin.c
@@ -696,7 +696,7 @@ modelica_real realMaxLit(void)
 
 modelica_integer intMaxLit(void)
 {
-  return LONG_MAX / 2;
+  return INT_MAX / 2;
 }
 
 modelica_boolean setStackOverflowSignal(modelica_boolean inSignal)
